requirements.txt
openai>=1.0.0
python-jobspy>=0.1.0
playwright>=1.40.0
httpx>=0.26.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
rich>=13.0.0
flask>=3.0.0
python-docx>=1.0.0
fpdf2>=2.7.0
pypdf>=4.0.0
pyyaml>=6.0.0
typer>=0.9.0
pydantic>=2.0.0
